updates on DTOs, mapping, and error handling

+++ budget/src/main/java/com/github/budget/service/SpecFileService.java
@@ -2,20 +2,22 @@ package com.github.budget.service;
 
 import java.io.File;
 import java.io.IOException;
+import java.util.Collections;
 import java.util.List;
 import java.util.Optional;
+import java.util.stream.Collectors;
 
 import org.bson.Document;
-import org.bson.types.ObjectId;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
-import com.fasterxml.jackson.core.exc.StreamReadException;
-import com.fasterxml.jackson.databind.DatabindException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.github.budget.constant.Constant;
+import com.github.budget.dto.request.SpecFileRequestDto;
+import com.github.budget.dto.response.SpecFileResponseDto;
 import com.github.budget.entity.SpecFile;
+import com.github.budget.mapper.SpecFileMapper;
 import com.github.budget.repository.SpecFileRepository;
 
 import lombok.AllArgsConstructor;
@@ -25,45 +27,40 @@ import lombok.AllArgsConstructor;
 public class SpecFileService {
     SpecFileRepository specFileRepository;
 
-    public ResponseEntity<List<SpecFile>> getSpecFiles() {
-        return ResponseEntity.ok(specFileRepository.findAll());
+    public List<SpecFileResponseDto> getSpecFiles() {
+        List<SpecFile> specFiles = specFileRepository.findAll();
+        return specFiles.stream()
+                .map(SpecFileMapper::mapToSpecFileDto)
+                .collect(Collectors.toList());
     }
 
-    public String createSpecFile(MultipartFile file)
+    public void createSpecFile(MultipartFile file)
             throws IllegalStateException, IOException {
 
         String destinationPath = Constant.FILE_DIR + file.getOriginalFilename();
         File destFile = new File(destinationPath);
 
-        // upload file to disk
-        file.transferTo(destFile);
-
         SpecFile specFile = new SpecFile();
         specFile.setFilename(file.getOriginalFilename());
         specFile.setFiletype(file.getContentType());
         specFile.setPath(destinationPath);
-        specFile.setSchema(JSONtoSpec(destinationPath));
+        // upload file to disk
+        file.transferTo(destFile);
+        // read file and convert to json
+        specFile.setSchema(fileJsonToSpec(destinationPath));
         specFileRepository.save(specFile);
-        return "Saved spec file to " + destinationPath + " successfully";
 
     }
 
-    public Document getSpecFileSchemaByFilename(String filename) {
-        Optional<SpecFile> specFile = specFileRepository.findByFilename(filename);
+    public List<SpecFileResponseDto> getSpecFilesByUsername(String username) {
 
-        if (specFile.isEmpty()) {
-            return null;
+        Optional<List<SpecFile>> specFiles = specFileRepository.findByCreatedBy(username);
+        if (specFiles.isEmpty()) {
+            return Collections.emptyList();
         }
-        return specFile.get().getSchema();
-    }
-
-    public ObjectId getSpecFileIdByFilename(String filename) {
-        Optional<SpecFile> specFile = specFileRepository.findByFilename(filename);
-
-        if (specFile.isEmpty()) {
-            return null;
-        }
-        return specFile.get().getId();
+        return specFiles.get().stream()
+                .map(SpecFileMapper::mapToSpecFileDto)
+                .collect(Collectors.toList());
     }
 
     public boolean deleteSpecFile(String filename) {
@@ -76,18 +73,36 @@ public class SpecFileService {
         File file = new File(spec.getPath());
 
         // deletes file from disk
-        file.delete();
+        if (!file.delete())
+            return false;
 
         specFileRepository.deleteByFilename(filename);
 
         return true;
     }
 
-    public Document JSONtoSpec(String filePath) throws StreamReadException, DatabindException, IOException {
+    public Document getSpecFileSchemaByFilename(String filename) {
+        Optional<SpecFile> specFile = specFileRepository.findByFilename(filename);
+
+        if (specFile.isEmpty()) {
+            return new Document();
+        }
+        return specFile.get().getSchema();
+    }
+
+    public String getSpecFileIdByFilename(String filename) {
+        Optional<SpecFile> specFile = specFileRepository.findByFilename(filename);
+
+        if (specFile.isEmpty()) {
+            return null;
+        }
+        return specFile.get().getId();
+    }
+
+    public Document fileJsonToSpec(String filePath) throws IOException {
         ObjectMapper objectMapper = new ObjectMapper();
 
-        Document spec = objectMapper.readValue(new File(filePath), Document.class);
-        return spec;
+        return objectMapper.readValue(new File(filePath), Document.class);
     }
 
 }
